Refactor NPC scripts into dedicated modules and unify shop & movement
