+++ Assets/Scripts/NPC/Combat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ecf369e7398c473893b9372894e9ed82
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
