+++ Assets/Scripts/NPC/Combat/BaseNpcCombat.cs
