+++ Assets/Scripts/NPC/Combat/BaseNpcCombat.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 65815a34833244dfad5b08855ea6ff7e
