+++ Assets/Scripts/NPC/Combat/NpcFacing.cs
