+++ Assets/Scripts/NPC/Combat/NpcFacing.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0ade83d186a2426484d785699fb0c95b
