+++ Assets/Scripts/NPC/Combat/NpcSpriteAnimator.cs
