+++ Assets/Scripts/NPC/Interaction.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 34df6e945f2449878a3dd5534da5a374
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
