+++ Assets/Scripts/NPC/Interaction/NpcInteractable.cs.meta
