+++ Assets/Scripts/NPC/Interaction/NpcShopOpener.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: adf97df505f4465792ad1b25e702c5df
