+++ Assets/Scripts/NPC/Movement.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 430466440c7e4c98a61e4af2f8cc63d4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
