+++ Assets/Scripts/NPC/Navigation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 403633e8a59c46218d0183e5198bdf93
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
