+++ Assets/Scripts/NPC/Navigation/WaypointPath.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace NPC
