+++ Assets/Scripts/NPC/Navigation/WaypointPath.cs.meta
