--- Assets/Scripts/NPC/NolofShopOpener.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d04fa88f0110fb24a8c34366b7af6b02
