--- Assets/Scripts/NPC/NpcRandomMovement.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: a0c533d25e324009bd37da6fc042d1ca
-timeCreated: 1755289570
