--- Assets/Scripts/NPC/ShopOpener.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: a1327c6bd3da4bcaad4c2ace865e9e6d
-timeCreated: 1755892739
