+++ Assets/Scripts/NPC/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eb0be8abcb6e404a954daff721f03d3f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
