+++ Assets/Scripts/NPC/UI/NpcLabel.cs.meta
